Moved global setup/teardown to test folder
