+++ jest.config.js
@@ -1,6 +1,6 @@
 // Jest configuration as per https://jestjs.io/docs/configuration
 
 module.exports = {
-    globalSetup: './test/end2end/__global-e2e-setup.js',
-    globalTeardown: './test/end2end/__global-e2e-teardown.js',
+    globalSetup: './test/__global-e2e-setup.js',
+    globalTeardown: './test/__global-e2e-teardown.js',
 }
