+++ test/__global-e2e-teardown.js
@@ -1,5 +1,5 @@
-const testEnv = require('./test-env');
-const resources = require('../resources');
+const testEnv = require('./end2end/test-env');
+const resources = require('./resources');
 
 // Global teardown for Jest.
 //
